requirements.txt
amazoncaptcha==0.5.11
beautifulsoup4==4.12.3
black==24.4.0
fastapi==0.110.1
Flask==3.0.3
google-cloud==0.34.0
google-cloud-pubsub==2.21.4
langchain==0.2.5
langchain-community==0.2.19
langchain-core==0.1.42
langchain-experimental==0.0.61
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
pandas==2.2.2
requests==2.32.2
selenium==4.19.0
uvicorn==0.25.0
webdriver-manager==4.0.1
